change things here and there to make it work
