+++ utils/timewiz/__init__.py
@@ -1,22 +1,3 @@
 from .game_phase import game_phase
 from .is_witching_hour import is_witching_hour
 from .timewiz import timewiz
-
-CW_OFFSET = -11_093_806_800
-CW3_OFFSET = -11_093_803_200
-SPEED = 3
-
-CW_MONTHS = ["Hailag", "Wintar", "Hornung", "Lenzin", "Ōstar", "Winni", "Brāh", "Hewi", "Aran", "Witu", "Wīndume", "Herbist", "Hailag", "Wintar"]
-CW_PERIODS = ["Night", "Morning", "Day", "Evening", "Night"]
-CW_SEASONS = ["Aries", "Taurus", "Gemini", "Cancer", "Leo", "Virgo", "Libra", "Scorpio", "Sagittarius", "Capricorn", "Aquarius", "Pisces", "Aries"]
-CW3_SEASONS = ["Овен", "Телец", "Близнецы", "Рак", "Лев", "Дева", "Весы", "Скорпион", "Стрелец", "Козерог", "Водолей", "Рыбы", "Овен"]
-CW_WEEKDAYS = ["Mânotag", "Ziestag", "Mittawehha", "Jhonarestag", "Frîatag", "Sunnûnabund", "Sunnûntag", "Mânotag"]
-CW3_WEEKDAYS = ["Понедельник", "Вторник", "Среда", "Четверг", "Пятница", "Суббота", "Воскресенье", "Понедельник"]
-HEB_MONTHS_ENG = ["Adar", "Nisan", "Iyyar", "Sivan", "Tammuz", "Av", "Elul", "Tishrei", "Marcheshvan", "Kislev", "Tevet", "Shevat", "Adar", "Adar"]
-HEB_MONTHS_VRT = ["אדר", "ניסן", "אייר", "סיוון", "תמוז", "אב", "אלול", "תשרי", "מרחשוון", "כסלו", "טבת", "שבט", "אדר", "אדר"]
-SEASON_EMOJI = "♈♉♊♋♌♍♎♏♐♑♒♓"
-CLOCK_EMOJI = "🕛🕧🕐🕜🕑🕝🕒🕞🕓🕟🕔🕠🕕🕡🕖🕢🕗🕣🕘🕤🕙🕥🕚🕦🕛"
-
-DEPRESSION = 8 + 3 / 5
-SUN_YAT_SEN_ERA = 2698
-WRITERS_ERA = 2637
