+++ utils/timewiz/timewiz.py
@@ -9,17 +9,24 @@ import astral
 from timezonefinder import TimezoneFinder
 
 import sesDate
+CW_OFFSET = -11_093_806_800
+CW3_OFFSET = -11_093_803_200
+SPEED = 3
 
-from . import CW_OFFSET, CW3_OFFSET
-from . import SPEED
-from . import CW_MONTHS, CW_PERIODS, CW_SEASONS, CW_WEEKDAYS
-from . import CW3_WEEKDAYS
-from . import HEB_MONTHS_ENG, HEB_MONTHS_VRT
-from . import SEASON_EMOJI, CLOCK_EMOJI
+CW_MONTHS = ["Hailag", "Wintar", "Hornung", "Lenzin", "Ōstar", "Winni", "Brāh", "Hewi", "Aran", "Witu", "Wīndume", "Herbist", "Hailag", "Wintar"]
+CW_PERIODS = ["Night", "Morning", "Day", "Evening", "Night"]
+CW_SEASONS = ["Aries", "Taurus", "Gemini", "Cancer", "Leo", "Virgo", "Libra", "Scorpio", "Sagittarius", "Capricorn", "Aquarius", "Pisces", "Aries"]
+CW3_SEASONS = ["Овен", "Телец", "Близнецы", "Рак", "Лев", "Дева", "Весы", "Скорпион", "Стрелец", "Козерог", "Водолей", "Рыбы", "Овен"]
+CW_WEEKDAYS = ["Mânotag", "Ziestag", "Mittawehha", "Jhonarestag", "Frîatag", "Sunnûnabund", "Sunnûntag", "Mânotag"]
+CW3_WEEKDAYS = ["Понедельник", "Вторник", "Среда", "Четверг", "Пятница", "Суббота", "Воскресенье", "Понедельник"]
+HEB_MONTHS_ENG = ["Adar", "Nisan", "Iyyar", "Sivan", "Tammuz", "Av", "Elul", "Tishrei", "Marcheshvan", "Kislev", "Tevet", "Shevat", "Adar", "Adar"]
+HEB_MONTHS_VRT = ["אדר", "ניסן", "אייר", "סיוון", "תמוז", "אב", "אלול", "תשרי", "מרחשוון", "כסלו", "טבת", "שבט", "אדר", "אדר"]
+SEASON_EMOJI = "♈♉♊♋♌♍♎♏♐♑♒♓"
+CLOCK_EMOJI = "🕛🕧🕐🕜🕑🕝🕒🕞🕓🕟🕔🕠🕕🕡🕖🕢🕗🕣🕘🕤🕙🕥🕚🕦🕛"
 
-from . import DEPRESSION  # depression is required to look at this code
-from . import SUN_YAT_SEN_ERA
-from . import WRITERS_ERA
+DEPRESSION = 8 + 3 / 5
+SUN_YAT_SEN_ERA = 2698
+WRITERS_ERA = 2637
 
 
 def timewiz(user_data):
